refactor(node): Stop sharing the dependency handler between NPM and Bun
The upcoming changes for implementing workspace support for NPM will
change the `NpmDependencyHandler`. With the recent introduction of `Bun`
NPM's dependency handler has become shared. Eliminate that sharing
again, to allow implementing workspace support for these two managers
independently.
